Change int_generator to produce long long instead
With int we would get an overflow when having a large
file size

+++ speed_test.cpp
@@ -43,7 +43,7 @@ int main(int argc, char ** argv) {
 
 	switch (cmd_options.item_type) {
 	case 0:
-		run_test<int_generator, file_stream<int>>();
+		run_test<int_generator, file_stream<long long>>();
 		break;
 	case 1:
 		run_test<string_generator, serialized_file_stream<std::string>>();
